requirements.txt
Django==5.2.7
scikit-learn==1.5.2
numpy==2.1.1
joblib==1.4.2
PyPDF2==3.0.1
python-docx==1.1.2
